new header and .cpp file
Moved functions related to matrix solving algorithm to another .cpp file and added a new header file. Make sure to include "matrix_solve.cpp" in your tasks file if you use this branch
