+++ .vscode/tasks.json
@@ -8,6 +8,7 @@
                 "-fdiagnostics-color=always",
                 "-g",
                 "parse.cpp",
+                "matrix_solve.cpp",
                 "${file}",
                 "-o",
                 "${fileDirname}\\${fileBasenameNoExtension}.exe"
